requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pg8000==1.31.5
python-dotenv==1.0.1
SQLAlchemy==2.0.37
alembic==1.14.0
